Port all the bag verifier tests to the TestCases model
