+++ bag_register/src/test/scala/uk/ac/wellcome/platform/archive/bag_register/fixtures/BagRegisterFixtures.scala
@@ -173,7 +173,7 @@ trait BagRegisterFixtures
         payloadFileCount = dataFileCount
       )
 
-    uploadBagObjects(bagObjects)
+    uploadBagObjects(bagRoot = bagRoot, objects = bagObjects)
 
     (bagRoot, bagInfo)
   }
