+++ bag_register/src/test/scala/uk/ac/wellcome/platform/archive/bag_register/services/RegisterTest.scala
@@ -151,7 +151,7 @@ class RegisterTest
         objLocation.copy(namespace = objLocation.namespace + "_wrong") -> contents
     }
 
-    uploadBagObjects(badBagObjects)
+    uploadBagObjects(bagRoot = bagRoot, objects = badBagObjects)
 
     val location = createPrimaryLocationWith(
       prefix = bagRoot
