+++ common/src/test/scala/uk/ac/wellcome/platform/archive/common/fixtures/s3/S3BagBuilder.scala
@@ -59,7 +59,7 @@ trait S3BagBuilder
     )
 
     implicit val typedStore: NewS3TypedStore[String] = NewS3TypedStore[String]
-    uploadBagObjects(bagObjects)
+    uploadBagObjects(bagRoot, objects = bagObjects)
 
     (bagRoot, bagInfo)
   }
